+++ src/Dfe.ManageSchoolImprovement/Pages/Contacts/Index.cshtml.cs
@@ -259,7 +259,7 @@ namespace Dfe.ManageSchoolImprovement.Frontend.Pages.Contacts
                             Email = SupportProject.SupportingOrganisationContactEmailAddress ?? "",
                             Phone = SupportProject.SupportingOrganisationContactPhone ?? "",
                             Address = SupportProject.SupportingOrganisationContactAddress ?? "",
-                            LastModifiedOn = SupportProject.DateSupportingOrganisationContactDetailsAdded
+                            LastModifiedOn = SupportProject.DateSupportOrganisationChosen
                         };
                     }
                 }
@@ -272,7 +272,7 @@ namespace Dfe.ManageSchoolImprovement.Frontend.Pages.Contacts
                         Email = SupportProject?.SupportingOrganisationContactEmailAddress ?? "",
                         Phone = SupportProject?.SupportingOrganisationContactPhone ?? "",
                         Address = SupportProject?.SupportingOrganisationContactAddress ?? "",
-                        LastModifiedOn = SupportProject?.DateSupportingOrganisationContactDetailsAdded
+                        LastModifiedOn = SupportProject?.DateSupportOrganisationChosen
                     };
                 }
             } 
@@ -285,7 +285,7 @@ namespace Dfe.ManageSchoolImprovement.Frontend.Pages.Contacts
                     Email = SupportProject.SupportingOrganisationContactEmailAddress ?? "",
                     Phone = SupportProject.SupportingOrganisationContactPhone ?? "",
                     Address = SupportProject.SupportingOrganisationContactAddress ?? "",
-                    LastModifiedOn = SupportProject.DateSupportingOrganisationContactDetailsAdded
+                    LastModifiedOn = SupportProject.DateSupportOrganisationChosen
                 };
             }
         }