+++ src/Dfe.ManageSchoolImprovement/Pages/TaskList/ChoosePreferredSupportingOrganisation/EnterSupportingOrganisationSchoolDetails.cshtml.cs
@@ -121,8 +121,7 @@ public class EnterSupportingOrganisationSchoolDetailsModel(
             SupportProject?.SupportingOrganisationContactName,
             SupportProject?.SupportingOrganisationContactEmailAddress,
             SupportProject?.SupportingOrganisationContactPhone,
-            SupportProject?.SupportingOrganisationContactAddress,
-            SupportProject?.DateSupportingOrganisationContactDetailsAdded);
+            SupportProject?.SupportingOrganisationContactAddress);
         
         var result = await mediator.Send(command, cancellationToken);